[symex] split builtin_functions.cpp into focused files
The 3201-line monolith is replaced by 10 files under
src/goto-symex/builtin_functions/, each covering one concern:
memory_alloc, io, cpp_memory, threads, va_arg, memory_ops,
object_size, run_builtin, python_builtins, misc. No logic
was changed; static linkage is preserved for all TU-local helpers.
