+++ src/goto-symex/builtin_functions/threads.cpp
@@ -1,0 +1,261 @@
+#include <cassert>
+#include <complex>
+#include <functional>
+#include <goto-symex/execution_state.h>
+#include <goto-symex/goto_symex.h>
+#include <goto-symex/reachability_tree.h>
+#include <goto-symex/printf_formatter.h>
+#include <iomanip>
+#include <limits>
+#include <sstream>
+#include <string>
+#include <util/arith_tools.h>
+#include <util/base_type.h>
+#include <util/c_types.h>
+#include <util/cprover_prefix.h>
+#include <util/expr_util.h>
+#include <util/i2string.h>
+#include <irep2/irep2.h>
+#include <util/message.h>
+#include <util/message/format.h>
+#include <util/migrate.h>
+#include <util/prefix.h>
+#include <util/std_types.h>
+#include <vector>
+#include <algorithm>
+#include <util/array2string.h>
+
+void goto_symext::intrinsic_yield(reachability_treet &art)
+{
+  // Don't context switch if the guard is false.
+  if (!cur_state->guard.is_false())
+    art.get_cur_state().force_cswitch();
+}
+
+void goto_symext::intrinsic_switch_to(
+  const code_function_call2t &call,
+  reachability_treet &art)
+{
+  // Switch to other thread.
+  const expr2tc &num = call.operands[0];
+  if (!is_constant_int2t(num))
+  {
+    log_error("Can't switch to non-constant thread id no\n{}", *num);
+    abort();
+  }
+
+  const constant_int2t &thread_num = to_constant_int2t(num);
+
+  unsigned int tid = thread_num.value.to_uint64();
+  if (tid != art.get_cur_state().get_active_state_number())
+    art.get_cur_state().switch_to_thread(tid);
+}
+
+void goto_symext::intrinsic_switch_from(reachability_treet &art)
+{
+  // Mark switching back to this thread as already having been explored
+  art.get_cur_state()
+    .DFS_traversed[art.get_cur_state().get_active_state_number()] = true;
+
+  // And force a context switch.
+  art.get_cur_state().force_cswitch();
+}
+
+void goto_symext::intrinsic_get_thread_id(
+  const code_function_call2t &call,
+  reachability_treet &art)
+{
+  statet &state = art.get_cur_state().get_active_state();
+
+  unsigned int thread_id = art.get_cur_state().get_active_state_number();
+
+  if (!is_nil_expr(call.ret))
+  {
+    expr2tc tid = constant_int2tc(call.ret->type, BigInt(thread_id));
+    state.value_set.assign(call.ret, tid);
+    symex_assign(code_assign2tc(call.ret, tid), true);
+  }
+}
+
+void goto_symext::intrinsic_set_thread_data(
+  const code_function_call2t &call,
+  reachability_treet &art)
+{
+  statet &state = art.get_cur_state().get_active_state();
+  expr2tc threadid = call.operands[0];
+  expr2tc startdata = call.operands[1];
+
+  // TODO: remove this global guard
+  state.global_guard.add(cur_state->guard.as_expr());
+  state.rename(threadid);
+  state.rename(startdata);
+
+  while (is_typecast2t(threadid))
+    threadid = to_typecast2t(threadid).from;
+
+  if (!is_constant_int2t(threadid))
+  {
+    log_error(
+      "__ESBMC_set_thread_internal_data received nonconstant thread id");
+    abort();
+  }
+  unsigned int tid = to_constant_int2t(threadid).value.to_uint64();
+  art.get_cur_state().set_thread_start_data(tid, startdata);
+}
+
+void goto_symext::intrinsic_get_thread_data(
+  const code_function_call2t &call,
+  reachability_treet &art)
+{
+  statet &state = art.get_cur_state().get_active_state();
+  expr2tc threadid = call.operands[0];
+
+  state.level2.rename(threadid);
+
+  while (is_typecast2t(threadid))
+    threadid = to_typecast2t(threadid).from;
+
+  if (!is_constant_int2t(threadid))
+  {
+    log_error(
+      "__ESBMC_get_thread_internal_data received nonconstant thread id");
+    abort();
+  }
+
+  unsigned int tid = to_constant_int2t(threadid).value.to_uint64();
+  const expr2tc &startdata = art.get_cur_state().get_thread_start_data(tid);
+
+  if (!is_nil_expr(call.ret))
+  {
+    assert(base_type_eq(call.ret->type, startdata->type, ns));
+
+    state.value_set.assign(call.ret, startdata);
+    symex_assign(code_assign2tc(call.ret, startdata), true);
+  }
+}
+
+void goto_symext::intrinsic_spawn_thread(
+  const code_function_call2t &call,
+  reachability_treet &art)
+{
+  if (
+    (k_induction || inductive_step) &&
+    !options.get_bool_option("disable-inductive-step"))
+  {
+    log_warning(
+      "k-induction does not support concurrency yet. Disabling inductive step");
+
+    // Disable inductive step on multi threaded code
+    options.set_option("disable-inductive-step", true);
+  }
+
+  // As an argument, we expect the address of a symbol.
+  expr2tc addr = call.operands[0];
+  simplify(addr); /* simplification is not needed for clang-11, but clang-13
+                   * inserts a no-op typecast here. */
+  assert(is_address_of2t(addr));
+  const address_of2t &addrof = to_address_of2t(addr);
+  assert(is_symbol2t(addrof.ptr_obj));
+  const irep_idt &symname = to_symbol2t(addrof.ptr_obj).thename;
+
+  goto_functionst::function_mapt::const_iterator it =
+    art.goto_functions.function_map.find(symname);
+  if (it == art.goto_functions.function_map.end())
+  {
+    log_error("Spawning thread \"{}\": symbol not found", symname);
+    abort();
+  }
+
+  if (!it->second.body_available)
+  {
+    log_error("Spawning thread \"{}\": no body", symname);
+    abort();
+  }
+
+  const goto_programt &prog = it->second.body;
+
+  // Invalidates current state reference!
+  unsigned int thread_id = art.get_cur_state().add_thread(&prog);
+
+  statet &state = art.get_cur_state().get_active_state();
+
+  expr2tc thread_id_exp = constant_int2tc(call.ret->type, BigInt(thread_id));
+
+  state.value_set.assign(call.ret, thread_id_exp);
+
+  symex_assign(code_assign2tc(call.ret, thread_id_exp), true);
+
+  // Force a context switch point. If the caller is in an atomic block, it'll be
+  // blocked, but a context switch will be forced when we exit the atomic block.
+  // Otherwise, this will cause the required context switch.
+  art.get_cur_state().force_cswitch();
+}
+
+void goto_symext::intrinsic_terminate_thread(reachability_treet &art)
+{
+  art.get_cur_state().end_thread();
+  // No need to force a context switch; an ended thread will cause the run to
+  // end and the switcher to be invoked.
+}
+
+void goto_symext::intrinsic_really_atomic_begin(reachability_treet &art)
+{
+  art.get_cur_state().increment_active_atomic_number();
+}
+
+void goto_symext::intrinsic_really_atomic_end(reachability_treet &art)
+{
+  art.get_cur_state().decrement_active_atomic_number();
+}
+
+void goto_symext::intrinsic_switch_to_monitor(reachability_treet &art)
+{
+  execution_statet &ex_state = art.get_cur_state();
+
+  // Don't do this if we're in the initialization function.
+  if (cur_state->source.pc->function == "__ESBMC_main")
+    return;
+
+  ex_state.switch_to_monitor();
+}
+
+void goto_symext::intrinsic_switch_from_monitor(reachability_treet &art)
+{
+  execution_statet &ex_state = art.get_cur_state();
+  ex_state.switch_away_from_monitor();
+}
+
+void goto_symext::intrinsic_register_monitor(
+  const code_function_call2t &call,
+  reachability_treet &art)
+{
+  execution_statet &ex_state = art.get_cur_state();
+
+  if (ex_state.tid_is_set)
+    assert(
+      0 && "Monitor thread ID was already set (__ESBMC_register_monitor)\n");
+
+  statet &state = art.get_cur_state().get_active_state();
+  expr2tc threadid = call.operands[0];
+  state.level2.rename(threadid);
+
+  while (is_typecast2t(threadid))
+    threadid = to_typecast2t(threadid).from;
+
+  if (!is_constant_int2t(threadid))
+  {
+    log_error("__ESBMC_register_monitor received nonconstant thread id");
+    abort();
+  }
+
+  unsigned int tid = to_constant_int2t(threadid).value.to_uint64();
+  assert(art.get_cur_state().threads_state.size() >= tid);
+  ex_state.monitor_tid = tid;
+  ex_state.tid_is_set = true;
+}
+
+void goto_symext::intrinsic_kill_monitor(reachability_treet &art)
+{
+  execution_statet &ex_state = art.get_cur_state();
+  ex_state.kill_monitor_thread();
+}
